[Port] Update paging APIs to use updated SCM types (#178)
In [System.ClientModel 1.1.0-beta.5](https://github.com/Azure/azure-sdk-for-net/blob/main/sdk/core/System.ClientModel/CHANGELOG.md#110-beta5-2024-07-11), `PageableCollection<T>` was changed to `PageCollection<T>` to better align convenience and protocol methods.

This PR upgrades the OpenAI library to SCM 1.1.0-beta.5, and manually implements internal types that will be generated from TypeSpec API specification in the future.  Specifically, it:

- Updates return types that have been renamed for streaming service methods
- For service methods corresponding to paginated-collection endpoints
  - Returns `PageCollection<T>` or `AsyncPageCollection<T>` from convenience methods
  - Takes `XxCollectionOptions` types that describe the paginated collection to be returned
  - Adds "rehydration" methods that return a collection that starts where another collection type left off via a persisted rehydration token
  - Returns `IEnumerable<ClientResult>` or `IAsyncEnumerable<ClientResult>` from protocol methods, representing a collection of service responses, where each corresponds to a page of results returned from the paginated endpoint
  - Implements internal page enumerator and page token types for each paginated-endpoint operation

Types that are intended to be output by the generator have been added in a temporary folder called `To.Be.Generated`.