Remove overloads that take complex parameters in favor of overloads that take simple string IDs (#203)
Taking a complex object when all that is actually needed is the object's `Id` property can be confusing. For example: New users might wonder if the values of the other properties can have side effects, and if they need to prepare the object in a certain way before passing it as an argument. By contrast, a plain string argument is much more straightforward. 

This change removes the overloads that take complex parameters in favor of the overloads that take simple string IDs. 

We plan to implement an alternative in the future that offers a similar kind of convenience but following a more natural OOP approach. For example, instead of having a method in the `AssistantClient` called `CancelRun` that takes a `ThreadRun` parameter, the `ThreadRun` class should be capable of acting like a client itself and have a method called `Cancel`. 

In other words, rather than having this:

```csharp
assistantClient.CancelRun(myRun);
```

We would like to have something like this instead:

```csharp
myRun.Cancel();
```